+++ .gitignore
@@ -1,2 +1,4 @@
 composer.lock
-vendor
+.idea
+vendor/
+bin/
