--- .travis.yml
@@ -1,14 +1,0 @@
-language: php
-
-php:
-  - 5.3
-  - 5.4
-  - 5.5
-
-before_script:
-  - composer self-update
-  - composer install --dev --prefer-source
-
-script:
-  - ./vendor/bin/phpunit --configuration tests/phpunit.xml
-  - sh -c "if [ '$TRAVIS_PHP_VERSION' != '5.3' ]; then ./vendor/bin/phpcs --standard=PSR2 ./src/; fi"
