--- autoload_classmap.php
@@ -1,18 +1,0 @@
-<?php
-// Generated by ZF2's ./bin/classmap_generator.php
-return array(
-    'SlmGoogleAnalytics\Module'                                     => __DIR__ . '/Module.php',
-    'SlmGoogleAnalytics\View\Helper\GoogleAnalytics'                => __DIR__ . '/src/SlmGoogleAnalytics/View/Helper/GoogleAnalytics.php',
-    'SlmGoogleAnalytics\Analytics\Ecommerce\Item'                   => __DIR__ . '/src/SlmGoogleAnalytics/Analytics/Ecommerce/Item.php',
-    'SlmGoogleAnalytics\Analytics\Ecommerce\Transaction'            => __DIR__ . '/src/SlmGoogleAnalytics/Analytics/Ecommerce/Transaction.php',
-    'SlmGoogleAnalytics\Analytics\Event'                            => __DIR__ . '/src/SlmGoogleAnalytics/Analytics/Event.php',
-    'SlmGoogleAnalytics\Analytics\Tracker'                          => __DIR__ . '/src/SlmGoogleAnalytics/Analytics/Tracker.php',
-    'SlmGoogleAnalytics\Exception\InvalidArgumentException'         => __DIR__ . '/src/SlmGoogleAnalytics/Exception/InvalidArgumentException.php',
-    'SlmGoogleAnalytics\Exception\RuntimeException'                 => __DIR__ . '/src/SlmGoogleAnalytics/Exception/RuntimeException.php',
-    'SlmGoogleAnalyticsTest\View\Helper\TestAsset\CustomViewHelper' => __DIR__ . '/tests/SlmGoogleAnalyticsTest/View/Helper/TestAsset/CustomViewHelper.php',
-    'SlmGoogleAnalyticsTest\View\Helper\GoogleAnalyticsTest'        => __DIR__ . '/tests/SlmGoogleAnalyticsTest/View/Helper/GoogleAnalyticsTest.php',
-    'SlmGoogleAnalyticsTest\Analytics\TrackerTest'                  => __DIR__ . '/tests/SlmGoogleAnalyticsTest/Analytics/TrackerTest.php',
-    'SlmGoogleAnalyticsTest\Analytics\EventTest'                    => __DIR__ . '/tests/SlmGoogleAnalyticsTest/Analytics/EventTest.php',
-    'SlmGoogleAnalyticsTest\Analytics\Ecommerce\ItemTest'           => __DIR__ . '/tests/SlmGoogleAnalyticsTest/Analytics/Ecommerce/ItemTest.php',
-    'SlmGoogleAnalyticsTest\Analytics\Ecommerce\TransactionTest'    => __DIR__ . '/tests/SlmGoogleAnalyticsTest/Analytics/Ecommerce/TransactionTest.php',
-);
