--- autoload_function.php
@@ -1,12 +1,0 @@
-<?php
-return function ($class) {
-    static $map;
-    if (!$map) {
-        $map = include __DIR__ . '/autoload_classmap.php';
-    }
-
-    if (!isset($map[$class])) {
-        return false;
-    }
-    return include $map[$class];
-};
