--- autoload_register.php
@@ -1,2 +1,0 @@
-<?php
-spl_autoload_register(include __DIR__ . '/autoload_function.php');
