+++ src/Exception/InvalidArgumentException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LaminasGoogleAnalytics\Exception;
+
+class InvalidArgumentException extends \InvalidArgumentException
+{
+
+}
