+++ src/Exception/RuntimeException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LaminasGoogleAnalytics\Exception;
+
+class RuntimeException extends \RuntimeException
+{
+
+}
