+++ src/Service/ScriptFactory.php
@@ -32,77 +32,35 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * @author      Jurian Sluiman <jurian@juriansluiman.nl>
+ * @author      Witold Wasiczko <witold@wasiczko.pl>
  * @copyright   2012-2013 Jurian Sluiman.
  * @license     http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @link        http://juriansluiman.nl
+ * @link        http://www.psd2html.pl
  */
-namespace SlmGoogleAnalytics\Analytics\Ecommerce;
+namespace LaminasGoogleAnalytics\Service;
 
-class Item
+use Interop\Container\ContainerInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
+use LaminasGoogleAnalytics\View\Helper\Script\ScriptInterface;
+
+/**
+ * Class ScriptFactory
+ *
+ * @package LaminasGoogleAnalytics\Service
+ */
+class ScriptFactory implements FactoryInterface
 {
-    protected $sku;
-    protected $price;
-    protected $quantity;
-    protected $product;
-    protected $category;
-
-    public function __construct($sku, $price, $quantity = null, $product = null, $category = null)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
-        $this->setSku($sku);
-        $this->setPrice($price);
-        $this->setQuantity($quantity);
-        $this->setProduct($product);
-        $this->setCategory($category);
-    }
+        $config     = $container->get('config');
+        $scriptName = $config['google_analytics']['script'];
 
-    public function getSku()
-    {
-        return $this->sku;
-    }
+        $script = $container->get($scriptName);
+        /* @var $script ScriptInterface */
+        $ga = $container->get('google-analytics');
 
-    public function setSku($sku)
-    {
-        $this->sku = $sku;
-    }
+        $script->setTracker($ga);
 
-    public function getProduct()
-    {
-        return $this->product;
-    }
-
-    public function setProduct($product)
-    {
-        $this->product = $product;
-    }
-
-    public function getCategory()
-    {
-        return $this->category;
-    }
-
-    public function setCategory($category)
-    {
-        $this->category = $category;
-    }
-
-    public function getPrice()
-    {
-        return $this->price;
-    }
-
-    public function setPrice($price)
-    {
-        $this->price = $price;
-    }
-
-    public function getQuantity()
-    {
-        return $this->quantity;
-    }
-
-    public function setQuantity($quantity)
-    {
-        $this->quantity = $quantity;
+        return $script;
     }
 }
