+++ src/Service/TrackerFactory.php
@@ -32,15 +32,47 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * @author      Jurian Sluiman <jurian@juriansluiman.nl>
+ * @author      Witold Wasiczko <witold@wasiczko.pl>
  * @copyright   2012-2013 Jurian Sluiman.
  * @license     http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @link        http://juriansluiman.nl
+ * @link        http://www.psd2html.pl
  */
 
-namespace SlmGoogleAnalytics\Exception;
+namespace LaminasGoogleAnalytics\Service;
 
-class InvalidArgumentException extends \InvalidArgumentException
+use Interop\Container\ContainerInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
+use LaminasGoogleAnalytics\Analytics\Tracker;
+
+final class TrackerFactory implements FactoryInterface
 {
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): Tracker
+    {
+        $config   = $container->get('config');
+        $gaConfig = $config['google_analytics'];
 
+        $tracker = new Tracker($gaConfig['id']);
+
+        if (isset($gaConfig['domain_name'])) {
+            $tracker->setDomainName($gaConfig['domain_name']);
+        }
+
+        if (isset($gaConfig['allow_linker'])) {
+            $tracker->setAllowLinker($gaConfig['allow_linker']);
+        }
+
+        if (true === $gaConfig['enable_display_advertising']) {
+            $tracker->setEnableDisplayAdvertising(true);
+        }
+
+        if (true === $gaConfig['anonymize_ip']) {
+            $tracker->setAnonymizeIp(true);
+        }
+
+        if (false === $gaConfig['enable']) {
+            $tracker->setEnableTracking(false);
+        }
+
+        return $tracker;
+    }
 }
