--- src/SlmGoogleAnalytics/Service/TrackerFactory.php
@@ -1,95 +1,0 @@
-<?php
-/**
- * Copyright (c) 2012-2013 Jurian Sluiman.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *
- *   * Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in
- *     the documentation and/or other materials provided with the
- *     distribution.
- *
- *   * Neither the names of the copyright holders nor the names of the
- *     contributors may be used to endorse or promote products derived
- *     from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- * @author      Witold Wasiczko <witold@wasiczko.pl>
- * @copyright   2012-2013 Jurian Sluiman.
- * @license     http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @link        http://www.psd2html.pl
- */
-namespace SlmGoogleAnalytics\Service;
-
-use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
-use SlmGoogleAnalytics\Analytics\Tracker;
-
-class TrackerFactory implements FactoryInterface
-{
-    /**
-     * @param ContainerInterface $container
-     * @param string             $requestedName
-     * @param array|null|null    $options
-     *
-     * @return Tracker
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
-    {
-        $config   = $container->get('config');
-        $gaConfig = $config['google_analytics'];
-
-        $tracker = new Tracker($gaConfig['id']);
-
-        if (isset($gaConfig['domain_name'])) {
-            $tracker->setDomainName($gaConfig['domain_name']);
-        }
-
-        if (isset($gaConfig['allow_linker'])) {
-            $tracker->setAllowLinker($gaConfig['allow_linker']);
-        }
-
-        if (true === $gaConfig['enable_display_advertising']) {
-            $tracker->setEnableDisplayAdvertising(true);
-        }
-
-        if (true === $gaConfig['anonymize_ip']) {
-            $tracker->setAnonymizeIp(true);
-        }
-
-        if (false === $gaConfig['enable']) {
-            $tracker->setEnableTracking(false);
-        }
-
-        return $tracker;
-    }
-
-    /**
-     * @param ServiceLocatorInterface $serviceLocator
-     *
-     * @return Tracker
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
-    {
-        return $this($serviceLocator, '');
-    }
-}
