+++ src/View/Helper/Script/ScriptInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaminasGoogleAnalytics\View\Helper\Script;
+
+use LaminasGoogleAnalytics\Analytics\Tracker;
+
+interface ScriptInterface
+{
+    public function setTracker(Tracker $tracker);
+    public function getCode();
+}
