+++ tests/phpunit.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <phpunit bootstrap="bootstrap.php" colors="true">
-    <testsuite name="SlmGoogleAnalytics">
+    <testsuite name="LaminasGoogleAnalytics">
         <directory>./</directory>
     </testsuite>
 </phpunit>
